Check Mesh Closed for 1, 2, 3, 4, 5.fbx

+++ FBXMeshBuilder.h
@@ -2,10 +2,6 @@
 #define _FBXMeshBuilder_h_
 
 #include "Common.h"
-#include <vector>
-#include <map>
-#include <assert.h>
-
 #include "Argument.h"
 #include "Mesh.h"
 
@@ -25,11 +21,11 @@ private:
 	bool BuildFbxMeshes(FbxScene* fbxScene, std::vector<FbxNode* >& fbxNodes, const std::vector<MeshArray>& meshArrays);
 	void BuildFbxMesh(FbxScene* fbxScene, FbxNode* fbxNode, const Mesh& mesh, const FbxString& name);
 
-	void FillColor(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, int ch);
-	void FillUV(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, int ch);
-	void FillNormal(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, int ch);
-	void FillTangent(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, int ch);
-	void FillBinormal(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, int ch);
+	void FillColor(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, size_t ch);
+	void FillUV(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, size_t ch);
+	void FillNormal(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, size_t ch);
+	void FillTangent(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, size_t ch);
+	void FillBinormal(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh, size_t ch);
 
 	void FillPolygon(bool useBatch, FbxMesh* dstMesh, const Mesh& mesh);
